refactor(nodebuilder/pruner): Move all migration logic to nodebuilder pruner module
